add script to publish documentation and optimize commit sizes
The install_docs.sh script generates the documentation and pushes
it onto the gh-pages branch on github, thus making it publicly
available.

By default Doxygen embeds the footer in every page and since it
contains the current date when the documentation was generated,
every file will change in the commit. Thus we add a CSS rule to
append the current date before the footer, and a set of scripts to
update p4c.css (since Doxygen itself does not care about filtering
other files than source).