print the max error message only if the error threshold is meaningful
If `maxErrorCount` happens to be set to 1 (as is the case if you want
to fail immediately) the exceeding errors threshold does not make much
sense. So we create another exception type and we print the message
only in cases where the `maxErrorCount` is larger than one (and the
`--maxErrorCount` option available).